pyproject.toml
[tool.poetry]
name = "ml-flow-eval"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.9.1"
langchain = "^0.2.16"
langchain-community = "^0.2.16"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
